pyproject.toml
[project]
name = "refresh-token-rotation"
version = "0.1.0"
description = "refresh token rotation recreate"
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "sqlalchemy (>=2.0.37,<3.0.0)",
    "fastapi (>=0.115.7,<0.116.0)",
    "pydantic-settings (>=2.7.1,<3.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "pydantic[email] (>=2.10.6,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "psycopg[binary] (>=3.2.4,<4.0.0)",
    "greenlet (>=3.1.1,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
ruff = "^0.9.3"

